requirements-dev.txt
-r requirements.txt
defusedxml
pytest
pytest-benchmark
pytest-mock
pytest-xdist
sphinx.testing